Fix SonarCloud findings (cycle 1)
Split aces_plan.py (559 lines) under the 500-line limit: frozen value objects + AcesPlanError move to aces_plan_types.py and ACL parsing to aces_acl.py, mirroring the aces_composition.py split. Replace the Any-typed composition builder with a PEP 695 constrained type parameter. Behavior unchanged; all names remain importable from aces_plan.
