+++ shifter/engine/provisioner/aces_plan.py
@@ -22,16 +22,18 @@ against the reference backend's private accessors.
 Sizing/image are exposed as ``None`` when the author omitted them, so the backend
 applies its own default (e.g. a GCE profile machine type) rather than a forced
 constant. It self-discriminates on the plan ``kind`` so an ``aces-range`` command
-run against a cyberscript ``range_config`` fails loudly.
+run against a cyberscript ``range_config`` fails loudly. The frozen value objects
+live in ``aces_plan_types`` and ACL parsing in ``aces_acl`` (Sonar file-size split);
+they are re-exported here so callers keep importing from ``aces_plan``.
 """
 
 from __future__ import annotations
 
 import re
-from collections.abc import Mapping
-from dataclasses import dataclass
+from collections.abc import Callable, Mapping
 from typing import Any
 
+from aces_acl import build_node_acls
 from aces_composition import (
     AcesPlanAccount,
     AcesPlanContent,
@@ -39,6 +41,14 @@ from aces_composition import (
     build_account,
     build_content,
     build_feature,
+)
+from aces_plan_types import (
+    AcesPlan,
+    AcesPlanAcl,
+    AcesPlanError,
+    AcesPlanImage,
+    AcesPlanNetwork,
+    AcesPlanNode,
 )
 
 __all__ = [
@@ -95,81 +105,6 @@ MINIMUM_ACES_SDL_VERSION = "0.19.1"
 MAXIMUM_ACES_SDL_VERSION_EXCLUSIVE = "0.20.0"
 
 _MIB = 1024 * 1024
-
-#: ACL vocabulary, mirrored from aces_backend_libvirt.acls (the reference backend).
-_ACL_ACTIONS = {"allow": "accept", "accept": "accept", "deny": "drop", "drop": "drop"}
-_ACL_WILDCARD_PROTOCOLS = frozenset({"", "all", "any"})
-_ACL_DIRECTIONS = frozenset({"in", "out", "inout"})
-
-
-class AcesPlanError(ValueError):
-    """Raised when a persisted range_config is not a well-formed serialized ACES plan."""
-
-
-@dataclass(frozen=True)
-class AcesPlanImage:
-    """Authored image reference (from ACES ``source``); resolved to a concrete
-    provider image by the backend at realization (ADR-032-R2)."""
-
-    name: str
-    version: str | None = None
-
-
-@dataclass(frozen=True)
-class AcesPlanAcl:
-    """A node's authored network ACL (mirror aces_backend_libvirt.acls fields).
-
-    ``action`` is normalized to ``accept``/``drop`` and ``protocol`` to
-    ``tcp``/``udp``/``all``; ``from_net``/``to_net`` are kept as the authored
-    network refs (resolved to concrete CIDRs at realization, fail-closed, so an
-    unresolvable *specified* endpoint is never widened into a broad allow).
-    """
-
-    name: str
-    action: str
-    direction: str
-    protocol: str
-    ports: tuple[int, ...]
-    from_net: str | None = None
-    to_net: str | None = None
-
-
-@dataclass(frozen=True)
-class AcesPlanNode:
-    """A compute node to provision, with authored intent extracted verbatim."""
-
-    address: str
-    name: str
-    os_family: str
-    count: int
-    network_addresses: tuple[str, ...]
-    ram_mib: int | None = None
-    vcpus: int | None = None
-    image: AcesPlanImage | None = None
-    acls: tuple[AcesPlanAcl, ...] = ()
-
-
-@dataclass(frozen=True)
-class AcesPlanNetwork:
-    """A network the range's nodes attach to."""
-
-    address: str
-    name: str
-    cidr: str | None = None
-    gateway: str | None = None
-    internal: bool = False
-
-
-@dataclass(frozen=True)
-class AcesPlan:
-    """The parsed serialized ACES plan: nodes + networks + composition for realization."""
-
-    aces_sdl_version: str
-    nodes: tuple[AcesPlanNode, ...]
-    networks: tuple[AcesPlanNetwork, ...]
-    content: tuple[AcesPlanContent, ...] = ()
-    accounts: tuple[AcesPlanAccount, ...] = ()
-    features: tuple[AcesPlanFeature, ...] = ()
 
 
 def _mapping(value: object) -> Mapping[str, Any]:
@@ -344,9 +279,12 @@ def _validate_versions(envelope: Mapping[str, Any]) -> str:
     return version
 
 
-def _build_composition(
-    builder: Any, resource_type: str, pairs: list[tuple[str, Mapping[str, Any]]], node_lookup: dict[str, str]
-) -> tuple[Any, ...]:
+def _build_composition[CompositionValue: (AcesPlanContent, AcesPlanAccount, AcesPlanFeature)](
+    builder: Callable[[Mapping[str, Any]], CompositionValue | None],
+    resource_type: str,
+    pairs: list[tuple[str, Mapping[str, Any]]],
+    node_lookup: dict[str, str],
+) -> tuple[CompositionValue, ...]:
     """Build every composition value object of one kind, failing closed (ADR-032-R7).
 
     A payload missing required fields (``build_*`` returns ``None``) is a malformed
@@ -354,7 +292,7 @@ def _build_composition(
     node is a dangling composition reference. Both abort before an ``AcesPlan`` is
     returned, so credential/content bootstrap can never bind to an absent node.
     """
-    built: list[Any] = []
+    built: list[CompositionValue] = []
     for address, payload in pairs:
         value = builder(payload)
         if value is None:
@@ -432,87 +370,6 @@ def parse_plan(range_config: dict[str, Any] | None) -> AcesPlan:
     )
 
 
-def _acl_str(value: object) -> str:
-    """Return ``value`` if it is a string, else an empty string."""
-    return value if isinstance(value, str) else ""
-
-
-def _acl_action(raw: Mapping[str, Any]) -> str:
-    """Normalize an ACL action to ``accept``/``drop`` (mirror the reference)."""
-    token = _acl_str(raw.get("action")).lower()
-    if not token:
-        raise AcesPlanError("ACL missing 'action'")
-    if token not in _ACL_ACTIONS:
-        raise AcesPlanError(f"ACL unknown action {token!r}")
-    return _ACL_ACTIONS[token]
-
-
-def _acl_protocol(raw: Mapping[str, Any]) -> str:
-    """Normalize an ACL protocol to ``tcp``/``udp``/``all`` (mirror the reference)."""
-    token = _acl_str(raw.get("protocol")).lower()
-    if token in {"tcp", "udp"}:
-        return token
-    if token in _ACL_WILDCARD_PROTOCOLS:
-        return "all"
-    raise AcesPlanError(f"ACL unknown protocol {token!r}")
-
-
-def _acl_direction(raw: Mapping[str, Any]) -> str:
-    """Return the ACL direction (``in``/``out``/``inout``); default ``inout``."""
-    token = _acl_str(raw.get("direction")).lower()
-    if not token:
-        return "inout"
-    if token not in _ACL_DIRECTIONS:
-        raise AcesPlanError(f"ACL unknown direction {token!r}")
-    return token
-
-
-def _acl_ports(raw: Mapping[str, Any]) -> tuple[int, ...]:
-    """Return the ACL's validated integer ports (1-65535), rejecting bools/invalid."""
-    raw_ports = raw.get("ports", ())
-    if not isinstance(raw_ports, list | tuple):
-        raise AcesPlanError("ACL 'ports' is not a list")
-    ports: list[int] = []
-    for port in raw_ports:
-        if not isinstance(port, int) or isinstance(port, bool) or not 0 < port <= 65535:
-            raise AcesPlanError(f"ACL invalid port {port!r}")
-        ports.append(port)
-    return tuple(ports)
-
-
-def _acl_ref(raw: Mapping[str, Any], key: str) -> str | None:
-    """Return the raw endpoint network ref (``from_net``/``to_net``); None if omitted."""
-    ref = _acl_str(raw.get(key)).strip()
-    return ref or None
-
-
-def _acl(raw: object, index: int) -> AcesPlanAcl:
-    """Build one AcesPlanAcl, failing loud (fail-closed) on any invalid field."""
-    if not isinstance(raw, Mapping):
-        raise AcesPlanError(f"ACL #{index} is not an object")
-    protocol = _acl_protocol(raw)
-    ports = _acl_ports(raw)
-    if ports and protocol == "all":
-        raise AcesPlanError(f"ACL #{index} ports require protocol 'tcp' or 'udp'")
-    return AcesPlanAcl(
-        name=_acl_str(raw.get("name")).strip() or f"acl-{index}",
-        action=_acl_action(raw),
-        direction=_acl_direction(raw),
-        protocol=protocol,
-        ports=ports,
-        from_net=_acl_ref(raw, "from_net"),
-        to_net=_acl_ref(raw, "to_net"),
-    )
-
-
-def _node_acls(payload: Mapping[str, Any]) -> tuple[AcesPlanAcl, ...]:
-    """Extract a node's authored ACLs from ``spec.infrastructure.acls``."""
-    raw_acls = _infrastructure_spec(payload).get("acls")
-    if not isinstance(raw_acls, list | tuple):
-        return ()
-    return tuple(_acl(raw, index) for index, raw in enumerate(raw_acls))
-
-
 def _node(address: str, payload: Mapping[str, Any], network_lookup: dict[str, str]) -> AcesPlanNode:
     """Build an AcesPlanNode, resolving network membership and ACL endpoints.
 
@@ -526,7 +383,7 @@ def _node(address: str, payload: Mapping[str, Any], network_lookup: dict[str, st
         if target is None:
             raise AcesPlanError(f"node {address} references unknown network {ref!r}")
         resolved.append(target)
-    acls = _node_acls(payload)
+    acls = build_node_acls(_infrastructure_spec(payload).get("acls"))
     for acl in acls:
         for endpoint in (acl.from_net, acl.to_net):
             if endpoint is not None and endpoint not in network_lookup:
